+++ execution/market.go
@@ -649,11 +649,15 @@ func (m *Market) GetID() string {
 func (m *Market) PostRestore(ctx context.Context) error {
 	m.settlement.Update(m.position.Positions())
 
-	pps := m.position.Parties()
 	if err := m.peggedOrders.ReconcileWithOrderBook(m.matching); err != nil {
 		return err
 	}
 
+	if err := m.liquidity.ReconcileWithOrderBook(m.matching); err != nil {
+		return err
+	}
+
+	pps := m.position.Parties()
 	peggedOrder := m.peggedOrders.GetAll()
 	parties := make(map[string]struct{}, len(pps)+len(peggedOrder))